ml-service/requirements.txt
flask
numpy
pillow
scikit-learn
joblib
opencv-python-headless

